highlight resource transfer messages
This simplifies my comms variable creation (which was all idempotent anyway) and allows me to pass "resourceType" as part of the payload to gui_chat.

My main concern here is that I was too precious with the previous implicit key based resourceType determination in gui_chat. Like am I the only caller and all of this code exists for no reason? Whatever it's fine.